Zone/requirements.txt
# Core ML & Computer Vision
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
facenet-pytorch>=2.5.3
insightface>=0.7.3
mtcnn>=0.1.1

# Face Detection & Processing
numpy>=1.24.0
Pillow>=10.0.0
scipy>=1.10.0

# Database & Vector Search
faiss-cpu>=1.7.4
scikit-learn>=1.3.0

# Web UI
gradio>=4.0.0
fastapi>=0.104.0
uvicorn>=0.24.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyyaml>=6.0

# Development & Testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Optional: GPU Support (uncomment if CUDA available)
# torch>=2.0.0+cu118 --index-url https://download.pytorch.org/whl/cu118
# torchvision>=0.15.0+cu118 --index-url https://download.pytorch.org/whl/cu118
# faiss-gpu>=1.7.4
